requirements.txt
absl-py==0.1.11
astor==0.6.2
bleach==1.5.0
bokeh==0.12.14
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloudpickle==0.5.2
cycler==0.10.0
cytoolz==0.9.0.1
dask==0.17.1
decorator==4.2.1
distributed==1.21.3
Flask==1.0.2
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
heapdict==1.0.0
html5lib==0.9999999
idna==2.6
imageio==2.2.0
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
locket==0.2.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
msgpack-python==0.5.5
munch==2.3.1
networkx==2.1
numpy==1.14.1
olefile==0.45.1
opencv-python==3.4.0.12
packaging==17.1
pandas==0.22.0
partd==0.3.8
Pillow==5.0.0
pretrainedmodels==0.6.0
protobuf==3.5.2
psutil==5.4.3
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
requests==2.18.4
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
seaborn==0.8.1
six==1.11.0
sortedcontainers==1.5.9
tblib==1.3.2
tensorboard==1.6.0
tensorboard-logger==0.1.0
tensorflow==1.6.0
termcolor==1.1.0
toolz==0.9.0
torch==0.4.0
torchfile==0.1.0
torchvision==0.2.1
tornado==5.0
tqdm==4.23.0
urllib3==1.22
visdom==0.1.7
Werkzeug==0.14.1
xlrd==1.1.0
zict==0.1.3
gunicorn==19.8.1